refactoring: extract core and osv
