+++ code/hsec-core/src/Security/Advisories/Core/HsecId.hs
@@ -1,4 +1,4 @@
-module Security.Advisories.HsecId
+module Security.Advisories.Core.HsecId
   (
     HsecId
   , hsecIdYear
