+++ code/hsec-tools/hsec-tools.cabal
@@ -33,37 +33,36 @@ library
   exposed-modules:
     Security.Advisories
     Security.Advisories.Convert.OSV
-    Security.Advisories.Definition
     Security.Advisories.Filesystem
     Security.Advisories.Generate.HTML
     Security.Advisories.Git
-    Security.Advisories.HsecId
     Security.Advisories.Parse
     Security.Advisories.Queries
-    Security.OSV
 
   build-depends:
-    , aeson                 >=2.0.1.0 && <3
-    , base                  >=4.14    && <4.20
-    , Cabal-syntax          >=3.8.1.0 && <3.11
+    , aeson                 >=2.0.1.0  && <3
+    , base                  >=4.14     && <4.20
+    , Cabal-syntax          >=3.8.1.0  && <3.11
     , commonmark            ^>=0.2.2
-    , commonmark-pandoc     >=0.2     && <0.3
-    , containers            >=0.6     && <0.7
+    , commonmark-pandoc     >=0.2      && <0.3
+    , containers            >=0.6      && <0.7
     , cvss
     , directory             <2
     , extra                 ^>=1.7.5
-    , filepath              >=1.4     && <1.5
+    , filepath              >=1.4      && <1.5
+    , hsec-core
     , lucid                 >=2.9.0
-    , mtl                   >=2.2     && <2.4
-    , pandoc-types          >=1.22    && <2
-    , parsec                >=3       && <4
+    , mtl                   >=2.2      && <2.4
+    , osv
+    , pandoc-types          >=1.22     && <2
+    , parsec                >=3        && <4
     , pathwalk              >=0.3
-    , process               >=1.6     && <1.7
+    , process               >=1.6      && <1.7
     , safe                  >=0.3
-    , text                  >=1.2     && <3
-    , time                  >=1.9     && <1.14
+    , text                  >=1.2      && <3
+    , time                  >=1.9      && <1.14
     , toml-parser           ^>=1.3.0.0
-    , validation-selective  >=0.1     && <1
+    , validation-selective  >=0.1      && <1
 
   hs-source-dirs:   src
   default-language: Haskell2010
@@ -86,6 +85,7 @@ executable hsec-tools
     , bytestring            >=0.10    && <0.13
     , Cabal-syntax          >=3.8.1.0 && <3.11
     , filepath              >=1.4     && <1.5
+    , hsec-core
     , hsec-tools
     , optparse-applicative  >=0.17    && <0.19
     , text                  >=1.2     && <3
@@ -108,6 +108,7 @@ test-suite spec
     , Cabal-syntax
     , cvss
     , directory
+    , hsec-core
     , hsec-tools
     , pretty-simple  <5
     , tasty          <1.5
