+++ code/hsec-tools/src/Security/Advisories.hs
@@ -1,10 +1,10 @@
 module Security.Advisories
-  ( module Security.Advisories.Definition
-  , module Security.Advisories.HsecId
+  ( module Security.Advisories.Core.Advisory
+  , module Security.Advisories.Core.HsecId
   , module Security.Advisories.Parse
   )
 where
 
-import Security.Advisories.Definition
-import Security.Advisories.HsecId
+import Security.Advisories.Core.Advisory
+import Security.Advisories.Core.HsecId
 import Security.Advisories.Parse
