+++ code/osv/README.md
@@ -1,0 +1,7 @@
+# osv
+
+This project aims to support [Open Source Vulnerability format](https://ossf.github.io/osv-schema/).
+
+## Building
+
+We aim to support both regular cabal-based and nix-based builds.
