+++ code/osv/osv.cabal
@@ -1,0 +1,59 @@
+cabal-version:      2.4
+name:               osv
+version:            0.1.0.0
+
+-- A short (one-line) description of the package.
+synopsis:
+  Open Source Vulnerability format
+
+-- A longer description of the package.
+description:
+  Open Source Vulnerability format.
+
+-- A URL where users can report bugs.
+-- bug-reports:
+
+-- The license under which the package is released.
+license:            BSD-3-Clause
+author:             David Christiansen
+maintainer:         david@davidchristiansen.dk
+
+-- A copyright notice.
+-- copyright:
+category:           Data
+extra-doc-files:    CHANGELOG.md
+
+tested-with:
+  GHC ==8.10.7 || ==9.0.2 || ==9.2.8 || ==9.4.8 || ==9.6.3 || ==9.8.1
+
+library
+  exposed-modules:
+    Security.OSV
+
+  build-depends:
+    , aeson                 >=2.0.1.0  && <3
+    , base                  >=4.14     && <4.20
+    , cvss
+    , text                  >=1.2      && <3
+    , time                  >=1.9      && <1.14
+
+  hs-source-dirs:   src
+  default-language: Haskell2010
+  ghc-options:
+    -Wall -Wcompat -Widentities -Wincomplete-record-updates
+    -Wincomplete-uni-patterns -Wpartial-fields -Wredundant-constraints
+
+test-suite spec
+  type:             exitcode-stdio-1.0
+  hs-source-dirs:   test
+  main-is:          Spec.hs
+  build-depends:
+    , base           <5
+    , osv
+    , tasty          <1.5
+    , tasty-hunit    <0.11
+
+  default-language: Haskell2010
+  ghc-options:
+    -Wall -Wcompat -Widentities -Wincomplete-record-updates
+    -Wincomplete-uni-patterns -Wpartial-fields -Wredundant-constraints
