+++ code/osv/test/Spec.hs
@@ -1,0 +1,11 @@
+{-# LANGUAGE OverloadedStrings #-}
+
+module Main where
+
+import Test.Tasty
+
+main :: IO ()
+main =
+    defaultMain $
+      testGroup "Tests"
+        []
